+++ src/selectors/TaskSelectors.js
@@ -18,6 +18,7 @@ import {
 import {
     getCategoryTaskSorters,
     getCombinedTaskFilterDefinitions,
+    getTaskColumnSorter,
     isShowCompletedTasks,
     isShowFutureTasks,
     isShowTaskHierarchy
@@ -74,8 +75,9 @@ export const getTasksFilteredBySelectedFilter = createSelector(
     getTaskFiltersFilteredByVisibleState,
     getCombinedTaskFilterDefinitions,
     getCategoryTaskSorters,
+    getTaskColumnSorter,
     isBusy,
-    (tasks, tasksMetaData, searchTaskValue, showTaskHierarchy, showCompletedTasks, showFutureTasks, selectedTaskFilter, selectedTaskFilterDate, taskFields, taskFilters, combinedTaskFilterDefinitions, categoryTaskSorters, busy) => {
+    (tasks, tasksMetaData, searchTaskValue, showTaskHierarchy, showCompletedTasks, showFutureTasks, selectedTaskFilter, selectedTaskFilterDate, taskFields, taskFilters, combinedTaskFilterDefinitions, categoryTaskSorters, taskColumnSorter, busy) => {
         if (busy) {
             return getTasksFilteredBySelectedFilterResult;
         }
@@ -128,7 +130,14 @@ export const getTasksFilteredBySelectedFilter = createSelector(
 
         filteredTasks.push(...parentsToAdd);
 
-        const result = sortObjects(filteredTasks, taskFields, selectedTaskFilter, store.getState(), getTasksMetaDataFilteredByVisibleState, showTaskHierarchy);
+        if (selectedTaskFilter && taskColumnSorter) {
+            selectedTaskFilter.sorters = [
+                taskColumnSorter,
+                ...(selectedTaskFilter.sorters || [])
+            ];
+        }
+
+        const result = sortObjects(filteredTasks, taskFields, selectedTaskFilter ? selectedTaskFilter.sorters : [], store.getState(), getTasksMetaDataFilteredByVisibleState, showTaskHierarchy);
         getTasksFilteredBySelectedFilterResult = result;
 
         return result;